Split pty_process into separate files for UNIX and Windows.
