+++ src/nvim/CMakeLists.txt
@@ -74,6 +74,9 @@ foreach(sfile ${NEOVIM_SOURCES})
   if(${f} MATCHES "^(regexp_nfa.c)$")
     list(APPEND to_remove ${sfile})
   endif()
+  if(WIN32 AND ${f} MATCHES "^(pty_process_unix.c)$")
+    list(APPEND to_remove ${sfile})
+  endif()
 endforeach()
 
 list(REMOVE_ITEM NEOVIM_SOURCES ${to_remove})
